Add files via upload
Occasionally the python script to find routes crashes, so there might be duplicates of a particular activity identifier in the list. As the list can be big your not going to see it, so this runs through the list and creates a text file that does not have any duplicates.